ipfixprobe - refactor OutputPlugin
